ingestion-service/requirements.txt
asyncssh>=2.14.0
boto3>=1.34.0
motor>=3.3.0
tenacity>=8.2.0
pydantic-settings>=2.1.0
pyyaml>=6.0
structlog>=24.1.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp>=1.23.0
prometheus-client>=0.20.0
aiohttp>=3.9.0
httpx>=0.26.0
